refactor: change name spaces and class name
